Undo dependency vendoring
While cargos algorithm isn't perfect, vendored dependencies
do have problems such as not being able to easily update them
and having to duplicate the compile time.

The original reason to vendor was because of the Rust 2018
changes, which broke a bunch of crates, however, since Azul
requires Rust 1.33 or higher now anyway (due to the use
of const fn), this is less of an issue now.